backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
python-dotenv
google-genai
httpx
tenacity
websockets
openai
pymupdf
sqlalchemy
passlib[bcrypt]
python-jose[cryptography]
pymysql
bcrypt==4.0.1
